Refactor: Extract flashReducer tests from reducers.spec.js
Closes #206
